backend/requirements.txt
pikepdf
bitstring
Flask
flask-cors
python-dotenv